Updating makefile
Makefile currently works for all the caches except for ben’s. I went
through and replaced the source code with the originals (bc I think
some of them had modifications).

+++ source/zzhong/cache.c
@@ -10,16 +10,16 @@
 cache_t create_cache(uint32_t maxmem, hash_func hash_customized, evict_func evict_customized){
     cache_t cache = (cache_t) malloc( sizeof(struct cache_obj));
     assert(cache!=NULL&&"cache fails to allocate");
-
+    
     // Initialize the cache
     cache->maxmem = maxmem;
-
+    
     // Initialize the linked_list pointed by the cache
     cache->linked_list = (linked_list_t) malloc( sizeof(struct linked_list_obj));
     assert(cache!=NULL&&"cache fails to allocate");
 
-    linked_list_initialize(cache->linked_list, evict_customized);
-
+    linked_list_initialize(cache->linked_list, evict_customized); 
+    
     // Initialize the hash_table pointed by the cache
     cache->hash_table = (hash_table_t) malloc( sizeof(struct hash_table_obj));
     assert(cache!=NULL&&"cache fails to allocate");
@@ -34,9 +34,9 @@ cache_t create_cache(uint32_t maxmem, hash_func hash_customized, evict_func evic
 // If maxmem capacity is exceeded, sufficient values will be removed
 // from the cache to accomodate the new value.
 void cache_set(cache_t cache, key_type key, val_type val, uint32_t val_size){
-
+    
     item_t item_to_set = hash_table_find_item(cache->hash_table, key);
-    // If key already exists, it will overwrite the old value.
+    // If key already exists, it will overwrite the old value.  
     if (item_to_set!=NULL){
         item_to_set->val = val;
         return;
@@ -51,13 +51,13 @@ void cache_set(cache_t cache, key_type key, val_type val, uint32_t val_size){
     if (cache->hash_table->current_size+1 > (uint32_t) ((double) cache->hash_table->buckets_num*LOAD_FACTOR)){
         resize_hash_table(cache->hash_table);
     }
-
-    item_t item = hash_table_set(cache->hash_table, key, val, val_size);
+    
+    item_t item = hash_table_set(cache->hash_table, key, val, val_size);   
     node_t node = linked_list_set(cache->linked_list);
 
     // //connect the item in hash_table with its corresponding node in linked list
     cache_connect_node_and_item(item,node);
-
+    
 }
 
 // Retrieve the value associated with key in the cache, or NULL if not found. It will move the item with key to the front of the linked_list
@@ -72,7 +72,7 @@ val_type cache_get(cache_t cache, key_type key, uint32_t * val_size){
     //If key is not found, assign 0 to val_size
     *val_size = 0;
 	return NULL;
-
+    
 }
 
 
@@ -81,11 +81,11 @@ void cache_delete(cache_t cache, key_type key){
         item_t item_ptr = hash_table_find_item(cache->hash_table, key);
         linked_list_delete(cache->linked_list, item_ptr->node_ptr);
         hash_table_delete(cache->hash_table, item_ptr);
-
+     
 }
 
 // Compute the total amount of memory used up by all cache values (not keys)
-uint64_t cache_space_used(cache_t cache){
+uint32_t cache_space_used(cache_t cache){
 
     return hash_table_space_used(cache->hash_table);
 
@@ -124,3 +124,4 @@ void draw_cache(cache_t cache){
     draw_hash_table(cache->hash_table);
 
 }
+
